apps/backend/pyproject.toml
[project]
name = "snowmind-backend"
version = "0.1.0"
description = "SnowMind — Autonomous AI yield optimizer backend"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.115.0",
    "uvicorn[standard]==0.32.0",
    "pydantic==2.9.2",
    "pydantic-settings==2.6.1",
    "python-multipart==0.0.17",
    "PuLP==2.9.0",
    "web3==7.4.0",
    "supabase==2.9.0",
    "APScheduler==3.10.4",
    "cryptography==43.0.3",
    "python-jose[cryptography]==3.3.0",
    "httpx>=0.26,<0.29",
    "slowapi==0.1.9",
]

[dependency-groups]
dev = [
    "pytest==8.3.3",
    "pytest-asyncio==0.24.0",
    "pytest-cov==6.0.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
